+++ tests/test_windows.py
@@ -1,12 +1,16 @@
-from frmb import read_hierarchy_from_root
+import pytest
+
+from frmb import read_menu_hierarchy
 from frmb._windows import generate_reg_from_hierarchy
+from frmb._windows import get_key_path_for_file_association
+from frmb._windows import get_file_association_for_key_path
 
 
 def test__generate_reg_from_hierarchy(data_dir):
 
     structure1_dir = data_dir / "structure1"
     structure1_studio_dir = structure1_dir / "studio"
-    hierarchy = read_hierarchy_from_root(structure1_studio_dir)
+    hierarchy = read_menu_hierarchy(structure1_studio_dir)
 
     reg_content = generate_reg_from_hierarchy(hierarchy)
     assert "\n".join(reg_content) != "", print("\n".join(reg_content))
@@ -15,7 +19,7 @@ def test__generate_reg_from_hierarchy(data_dir):
 def test__generate_reg_from_hierarchy__comments(data_dir):
     structure1_dir = data_dir / "structure1"
     structure1_studio_dir = structure1_dir / "studio"
-    hierarchy = read_hierarchy_from_root(structure1_studio_dir)
+    hierarchy = read_menu_hierarchy(structure1_studio_dir)
 
     header_comments = [";;some comment", "; wow much fun"]
     reg_content = generate_reg_from_hierarchy(
@@ -30,3 +34,78 @@ def test__generate_reg_from_hierarchy__comments(data_dir):
     )
     result = [line for line in reg_content if line.lstrip("; ") in header_comments]
     assert len(result) == len(header_comments)
+
+
+def test__get_key_path_for_file_association():
+    expected = "HKEY_CURRENT_USER\\Software\\Classes\\*\\shell"
+    result = get_key_path_for_file_association("*")
+    assert result == expected
+
+    expected = "HKEY_LOCAL_MACHINE\\Software\\Classes\\*\\shell"
+    result = get_key_path_for_file_association("*", False)
+    assert result == expected
+
+    expected = "HKEY_LOCAL_MACHINE\\Software\\Classes\\Directory\\shell"
+    result = get_key_path_for_file_association("directory", False)
+    assert result == expected
+
+    expected = (
+        "HKEY_LOCAL_MACHINE\\Software\\Classes\\SystemFileAssociations\\.abc\\shell"
+    )
+    result = get_key_path_for_file_association(".abc", False)
+    assert result == expected
+
+    with pytest.raises(ValueError):
+        get_key_path_for_file_association("abc", False)
+
+    expected = "HKEY_CURRENT_USER\\Software\\Classes\\Drive\\shell"
+    result = get_key_path_for_file_association("drive", True)
+    assert result == expected
+
+    expected = "HKEY_LOCAL_MACHINE\\Software\\Classes\\Directory\\Background\\shell"
+    result = get_key_path_for_file_association("directory_background", False)
+    assert result == expected
+
+
+def test__get_file_association_for_key_path():
+    source = "HKEY_CURRENT_USER\\Software\\Classes\\*\\shell"
+    expected = ("*", True)
+    result = get_file_association_for_key_path(source)
+    assert result == expected
+
+    source = "HKEY_LOCAL_MACHINE\\Software\\Classes\\*\\shell"
+    expected = ("*", False)
+    result = get_file_association_for_key_path(source)
+    assert result == expected
+
+    source = "HKEY_LOCAL_MACHINE\\Software\\Classes\\Directory\\shell"
+    expected = ("directory", False)
+    result = get_file_association_for_key_path(source)
+    assert result == expected
+
+    source = (
+        "HKEY_LOCAL_MACHINE\\Software\\Classes\\SystemFileAssociations\\.abc\\shell"
+    )
+    expected = (".abc", False)
+    result = get_file_association_for_key_path(source)
+    assert result == expected
+
+    with pytest.raises(ValueError):
+        source = (
+            "HKEY_LOCAL_MACHINE\\Software\\Classes\\SystemFileAssociations\\abc\\shell"
+        )
+        get_file_association_for_key_path(source)
+
+    with pytest.raises(ValueError):
+        source = "HKEY_LOCAL_MACHINE\\Foo\\Classes\\SystemFileAssociations\\.abc\\shell"
+        get_file_association_for_key_path(source)
+
+    source = "HKEY_CURRENT_USER\\Software\\Classes\\Drive\\shell"
+    expected = ("drive", True)
+    result = get_file_association_for_key_path(source)
+    assert result == expected
+
+    source = "HKEY_LOCAL_MACHINE\\Software\\Classes\\Directory\\Background\\shell"
+    expected = ("directory_background", False)
+    result = get_file_association_for_key_path(source)
+    assert result == expected